requirements.txt
accelerate==1.5.1
async-timeout==4.0.3
bitsandbytes==0.42.0
einops==0.8.1
faiss-cpu==1.10.0
huggingface-hub==0.29.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.44
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.3.13
llama-index-embeddings-huggingface==0.5.2
llama-index-llms-huggingface==0.4.2
llama-index-vector-stores-faiss==0.3.0
llama-index-readers-file>=0.3.1
loguru==0.5.3
orjson==3.10.15
psutil==7.0.0
requests-toolbelt==1.0.0
safetensors==0.5.3
sentence-transformers==3.4.1
sympy==1.13.1
text-generation==0.7.0
tokenizers==0.21.0
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
transformers==4.49.0
zstandard==0.23.0
